Added a TrabantTouch file server used for synchronizing source with device. Also added:  - support for starting a member thread with an argument,  - support for fetching documents directory in SystemManager.

+++ UiTBC/Source/UiOpenGLRenderer.cxx
@@ -1272,7 +1272,7 @@ unsigned OpenGLRenderer::RenderScene()
 		if (GetFogFar() > 0)
 		{
 			::glEnable(GL_FOG);
-			::glFogf(GL_FOG_MODE, (GetFogExponent()>0)? GL_EXP : GL_LINEAR);
+			::glFogf(GL_FOG_MODE, (GetFogExponent()>0)? (float)GL_EXP : (float)GL_LINEAR);
 			::glFogf(GL_FOG_DENSITY, GetFogDensity());
 			::glFogf(GL_FOG_START, GetFogNear());
 			::glFogf(GL_FOG_END, GetFogFar());
